% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2.R
\name{aggregate_scans}
\alias{aggregate_scans}
\title{Aggregate Scans}
\usage{
aggregate_scans(
  group_scans,
  MS2tol,
  n_top_spectra_summed = 3L,
  quality_weights = c(purity = 2, quality = 1)
)
}
\arguments{
\item{group_scans}{a tibble which minimally contains precursorMz, precursorPurity, quality (derived from the parent peak), and data.}

\item{MS2tol}{a list of MS2 tol parameters produced with \code{\link{build_clamr_config}}}

\item{n_top_spectra_summed}{integer counts of maximum number of spectra to aggregate}

\item{quality_weights}{length 2 named vector with names "purity" and "quality" indicating the relative amount to weight by precursor purity (i.e., the amount of isolated signal matching the precursorMz) versus peak quality (i.e., good peak shapes).}
}
\description{
Combine multiple scans into a consensus spectrum
}
