% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adducts.R
\name{apply_coelution_labels}
\alias{apply_coelution_labels}
\title{Apply Coelution Labels}
\usage{
apply_coelution_labels(coelutions, coelution_labels)
}
\arguments{
\item{coelutions}{A list containing coelutions and coelution_mz_diffs of the \code{coelution} class produced by \code{find_common_coelutions}}

\item{coelution_labels}{a data_frame containing:
\itemize{
  \item{mz_set}
  \item{mz_subset}
  \item{library_formula - a formula label}
  \item{library_is_loss - TRUE/FALSE - whether the coelution difference is due to a gain or loss of the monoisotopic / non-adduct form}
  \item{library_type - label to apply to the connections - valid labels are "adduct", "isotopologue" or "distinct compound"}
  }}
}
\value{
a coelution object with labelled mz differences (added variables: library_formula, library_is_loss, library_type)
}
\description{
Apply Coelution Labels
}
