% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_searches.R
\name{augment_peakgroups}
\alias{augment_peakgroups}
\title{Augment Peak Groups}
\usage{
augment_peakgroups(mzroll_db_con)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}
}
\description{
Extract peakgroups table and add peakgroup level summaries from peaks
}
