% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_clamr_config.R
\name{build_clamr_config}
\alias{build_clamr_config}
\title{Build CLaM R config}
\usage{
build_clamr_config(param_list, quietly = FALSE)
}
\arguments{
\item{param_list}{A named list of parameters}

\item{quietly}{Hide messages and warnings}
}
\value{
a named list of mass spec parameters
}
\description{
Build CLaM R config
}
