% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_clamr_config.R
\name{clamr_config_assign}
\alias{clamr_config_assign}
\title{CLaM R Config Assign}
\usage{
clamr_config_assign(clamr_config)
}
\arguments{
\item{clamr_config}{a named list of mass spec parameters with special formatting of instrument tolerances generated by \code{\link{build_clamr_config}}.}
}
\description{
Write each entry of the config to objects (for debugging)
}
