% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakgrp_coelution_annotation.R
\name{combine_sample_coelutions_w_peaks}
\alias{combine_sample_coelutions_w_peaks}
\title{Combine Sample Coelutions with Peaks}
\usage{
combine_sample_coelutions_w_peaks(sample_peaks, sample_coelutions, MS1tol)
}
\arguments{
\item{sample_peaks}{peaks from a single sample from .mzrollDB}

\item{sample_coelutions}{coelutions from a single sample from .mzrollDB}

\item{MS1tol}{from \code{\link{build_clamr_config}} or \code{\link{format_mass_accuracy_input}}}
}
\description{
For a single sample, from coelution mz pairs (essentially edges) infer the peaks [mz, scans] then join these coelution peaks with peaks in the clamr_db.
}
\details{
The edge-list defines the connection between pairs of mzs over a scan interval.
In order to build a graph from these connections, we need to be able to define repeated observation of the same node in terms of a common mz and scan.
}
