% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry_utils.R
\name{create_isotopic_chemical_formula}
\alias{create_isotopic_chemical_formula}
\alias{parse_isotopic_chemical_formula}
\title{Create chemical formula}
\usage{
create_isotopic_chemical_formula(labeled_counts)

parse_isotopic_chemical_formula(formula)
}
\arguments{
\item{labeled_counts}{a tibble (or data.frame) containing two variables which will be used: label and count}

\item{formula}{a chemical formula which may included losses as generated by \code{\link{create_isotopic_chemical_formula}}}
}
\value{
a length one character vector containing a chemical formula
Explicitly check number of rows in mass_gains and mass_losses before
returning.
}
\description{
Read an internal, non-standard formula convention which allows for gains, losses and isotopes
}
\examples{
parse_isotopic_chemical_formula("C(6) H(12) O(6)")
parse_isotopic_chemical_formula("13C C(5) H(12) O(6)")
parse_isotopic_chemical_formula("13C C(5) H(12) O(6) - H(2)")

}
