% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaksets.R
\name{create_peakset}
\alias{create_peakset}
\title{Create Peakset
  (which may be missing some features)}
\usage{
create_peakset(peakgroups, peaks)
}
\arguments{
\item{peakgroups}{a table of unique peakgroups}

\item{peaks}{a table of sample x peakgroup peak abundances}
}
\value{
a peakset
}
\description{
Create Peakset
  (which may be missing some features)
}
\examples{
library(dplyr)

peakgroups <- tibble::tibble(groupId = 1:10)
samples <- tibble::tibble(sampleId = 1:10)
peaks <- tidyr::crossing(peakgroups, samples) \%>\%
  dplyr::mutate(peakId = 1:dplyr::n())

create_peakset(peakgroups, peaks)

}
