% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamr_data.R
\docType{data}
\name{elemental_frequency_summary}
\alias{elemental_frequency_summary}
\title{Table of isotopic information, and relative abundance in various media}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 9 columns.
}
\usage{
elemental_frequency_summary
}
\description{
Table of isotopic information, and relative abundance in various media
}
\keyword{data}
