% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{estimate_mz_ppm_drift}
\alias{estimate_mz_ppm_drift}
\title{Estimate M/Z ppm Drift}
\usage{
estimate_mz_ppm_drift(sample_matched_compounds, return_plot_data = FALSE)
}
\arguments{
\item{sample_matched_compounds}{tibble containing sample compound_id and rt}

\item{return_plot_data}{return data which can be used to plot alignment summaries with \code{\link{ms2_driven_aligner_plotting}}}
}
\value{
a list containing sample specific gam fits and a sample independent fit which should be used to correct sample-level overfitting
}
\description{
Determine the relative mass accuracy drift among experimental samples in ppm.
}
\details{
Alternate between estimating compound-specific retention time and sample-specific retention time deviations.
}
