% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_mzroll_db_con.R
\name{expand_with_mzroll_defaults}
\alias{expand_with_mzroll_defaults}
\title{Expand With Mzroll Defaults}
\usage{
expand_with_mzroll_defaults(mzroll_data)
}
\arguments{
\item{mzroll_data}{a list of tables corresponding to an mzrollDB SQL
schema, but possibly missing some variables.}
}
\value{
mzroll_data with default values added for variables which were
  not initially included
}
\description{
Expand With Mzroll Defaults
}
\examples{
library(dplyr)

peakgroups <- tibble::tibble(groupId = 1:10, compoundName = LETTERS[1:10])
samples <- tibble::tibble(sampleId = 1:10)
peaks <- tidyr::crossing(peakgroups, samples) \%>\%
  dplyr::mutate(peakId = 1:dplyr::n())

mzroll_data <- list(
  peakgroups = peakgroups,
  samples = samples,
  peaks = peaks
)

expand_with_mzroll_defaults(mzroll_data)

}
