% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2.R
\name{extract_and_group_fragments}
\alias{extract_and_group_fragments}
\title{Extract and group fragments}
\usage{
extract_and_group_fragments(mz_set, MS2tol, inconsistency_proportion = 0.005)
}
\arguments{
\item{mz_set}{a set of peaks with a consistent parent mass m/z which may contain multiple species with distinct RT / MS2 fingerprints.}

\item{MS2tol}{a list of MS2 tol parameters produced with \code{\link{build_clamr_config}}}

\item{inconsistency_proportion}{fraction of signal which can exist between peaks to still allow them to be resolved}
}
\value{
a tibble of an aggregated MS2 spectra
}
\description{
For a set of MS2 spectra pull out all MS2 fragments and group them into sets by fragment m/z
}
