% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2.R
\name{extract_fragments}
\alias{extract_fragments}
\title{Extract fragments only}
\usage{
extract_fragments(ms2scan)
}
\arguments{
\item{ms2scan}{a single ms2 scan, as produced by nested_peakgroup_features$scans[[x]]}
}
\description{
Read out the ms2 data exactly as it was captured from the instrument.
}
