% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaksets.R
\name{extract_peakset}
\alias{extract_peakset}
\title{Extract Peakset}
\usage{
extract_peakset(mzroll_db_con, searchTable = NULL)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}

\item{searchTable}{character vector of searchTables to use (or NULL for all)}
}
\value{
a peakset object (list of peaks and peakgroups)
}
\description{
Extract Peakset
}
