% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_similarity.R
\name{filter_low_complexity_spectra}
\alias{filter_low_complexity_spectra}
\title{Filter Low Complexity Spectra}
\usage{
filter_low_complexity_spectra(
  grouped_ms2_data,
  min_signal_fragments = 5,
  signal_fraction_co = 0.01
)
}
\arguments{
\item{grouped_ms2_data}{output of \code{\link{extract_and_group_fragments}}}

\item{min_signal_fragments}{minimum number of fragments that a spectra must have to be usable}

\item{signal_fraction_co}{fraction of total signal that a fragment must represent to be retained when using "signal_pool".}
}
\value{
grouped_ms2_data with low-complexity spectra and their fragments removed.
}
\description{
Filter peaks and their associated spectra based on the number of fragments accounting for a fraction of the signal (i.e., > 5 fragments each account for more than 1% of the signal)
}
