% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_similarity.R
\name{filter_low_information_fragments}
\alias{filter_low_information_fragments}
\title{Filter Low Information Fragments}
\usage{
filter_low_information_fragments(
  grouped_ms2_data,
  filter_method = "top5union",
  signal_fraction_co = 0.001
)
}
\arguments{
\item{grouped_ms2_data}{output of \code{\link{extract_and_group_fragments}}}

\item{filter_method}{method for filtering low-information fragments:
\describe{
  \item{top5union (default)}{the union of the 5 most abundant fragments (w/ intensity >1\% of total signal)}
  \item{frequency}{take fragments which are present in more than 10 spectra or >= 20\% of spectra or that are the most abundant ion of any spectra.}
  \item{signal_pool}{take fragments which represents more than \code{signal_fraction_co} fraction of the total signal across all peak spectra.}
}}

\item{signal_fraction_co}{fraction of total signal that a fragment must represent to be retained when using "signal_pool".}
}
\value{
grouped_ms2_data with low-information fragments removed.
}
\description{
Filter to high information fragments based on one of several filter_methods
}
