% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_anchors.R
\name{find_candidate_anchors}
\alias{find_candidate_anchors}
\title{Find Candidate Anchor Points}
\usage{
find_candidate_anchors(
  mzroll_db_con,
  clamr_config,
  n_anchors = 30L,
  anchor_rank_cutoff = 0.5
)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}

\item{clamr_config}{a named list of mass spec parameters with special formatting of instrument tolerances generated by \code{\link{build_clamr_config}}.}

\item{n_anchors}{number of anchor points to search for.}

\item{anchor_rank_cutoff}{rank cutoff for a viable anchors, smaller is better, ranks are scaled to [0,1].}
}
\value{
a tibble containing the 10-best scoring {mz, rt} pairs within each rt bin.
}
\description{
Identify peaks which are exclusively associated with a given m/z and thus
can be unambiguously identified across samples for retention time alignment.
}
\details{
The function aims to identify {mz, rt} pairs which can serve as anchor points
for retention time alignment. MAVEN will find these anchors in each sample
and enforce that all anchors peaks have the same retention time, other peaks
retention times will then be shifted by linearly interpolating between
flanking anchors.
}
\examples{
\dontrun{
future::plan("multicore")
mzroll_db_con <- clamr::mzroll_db_sqlite("/tmp/M005A-set123.mzrollDB")
clamr_config <- clamr::build_clamr_config(list(MS1tol = "10ppm", MS2tol = "20ppm"))
find_candidate_anchors(mzroll_db_con, clamr_config, n_anchors = 30L, anchor_rank_cutoff = 0.5)
}

}
