% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adducts.R
\name{find_common_coelutions}
\alias{find_common_coelutions}
\title{Find Common Coelutions}
\usage{
find_common_coelutions(adduct_path, coelution_n_cutoff = 0L)
}
\arguments{
\item{adduct_path}{A file containing a set of pairs of EICs (across scan_window_length scans) with correlated intensities across scans.}

\item{coelution_n_cutoff}{Minimum membership size of a cluster of mass differences to retain.}
}
\description{
Among a set of coeluting peaks, identify common mass differences as an indication of adducts or isotopologues
}
\details{
This function reads in an adduct file which contains candidate coelutions for a sample. These are generated using mzDeltas.cpp.
}
