% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_similarity.R
\name{find_consistent_ms2_fingerprints}
\alias{find_consistent_ms2_fingerprints}
\title{Find Consistent MS2 Fingerprints}
\usage{
find_consistent_ms2_fingerprints(
  mz_set,
  MS2tol,
  cosine_cutoff = 0.95,
  create_plots = FALSE,
  rt_range = NULL
)
}
\arguments{
\item{mz_set}{a set of peaks with a consistent parent mass m/z which may contain multiple species with distinct RT / MS2 fingerprints.}

\item{MS2tol}{mass tolerance for grouping ms2 m/z values: see \code{\link{build_clamr_config}}.}

\item{cosine_cutoff}{Minimum cosine similarity between a pair of MS2 fragment profiles to group them into a common cluster.}

\item{create_plots}{TRUE/FALSE: should a summary plot be printed.}

\item{rt_range}{optional if create_plots is FALSE, if create_plots is TRUE, length two numeric vector specifying the lower and upper limits of retention time for an experiment.}
}
\value{
assignments of parent peaks to distinct ms2 peak groups
}
\description{
Compare all MS2 fingerprints within the group to separate distinct groups based on MS2 profiles.
}
\examples{
MS2tol <- build_clamr_config(list(MS2tol = "20ppm"))$MS2tol
find_consistent_ms2_fingerprints(
  mz_set = clamr::mz_set_example,
  MS2tol = MS2tol, create_plots = TRUE, rt_range = c(0, 22)
)

}
