% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_utils.R
\name{find_density_minima}
\alias{find_density_minima}
\title{Separate by local minima}
\usage{
find_density_minima(x, bw = NULL, print_plot = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{bw}{the sd of the KDE smoother}

\item{print_plot}{TRUE/FALSE print a plot showing}
}
\value{
a vector of integer group labels
}
\description{
Determine the density of x (using the density function from stats) and then bin all points x using local minima as boundaries
}
