% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{find_mz_jumps}
\alias{find_mz_jumps}
\title{Find M/Z Jumps}
\usage{
find_mz_jumps(
  sorted_mzs,
  mass_accuracy_tolerance,
  absolute_or_relative,
  n_max_gap_inconsistencies = 0L,
  collapse = TRUE
)
}
\arguments{
\item{sorted_mzs}{numeric vector of m/z values}

\item{mass_accuracy_tolerance}{\itemize{
  \item{If \code{absolute_or_relative} == "relative": minimum fractional mass tolerance (m/z * tol) between consecutive m/z values that indicates to start a new M/Z group}
  \item{If \code{absolute_or_relative} == "absolute": minimum mass tolerance (Da) between consecutive m/z values that indicates to start a new M/Z group}
}}

\item{absolute_or_relative}{Indicate whether the specified \code{mass_accuracy_tolerance} is "relative" or "absolute"}

\item{n_max_gap_inconsistencies}{Maximum number of observations of an mz-value to ignore when find breaks between common mzs (i.e., group A & B are seperate if there are <= n_max_gap_inconsistencies mzs within the tolerance between them)}

\item{collapse}{\itemize{
  \item{If \code{collapse} == TRUE return mappings between each unique sorted_mz and mz_set}
  \item{If \code{collapse} == FALSE return a vector of mz_sets aligned to the sorted_mzs}
}}
}
\value{
a tibble containing paired mz and mz_set (mz groupings)
}
\description{
Seperate an ordered vector of M/Z value into distinct groups with similar M/Z by finding gaps of greater than \code{cutoff}
}
\seealso{
\code{\link{find_density_minima}}
}
