% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{find_scan_jumps}
\alias{find_scan_jumps}
\title{Returns a vector of grouped scans -- each group is defined by a separation of less than tolerances}
\usage{
find_scan_jumps(scans, tolerances)
}
\arguments{
\item{scans}{a vector of scan numbers}

\item{tolerances}{tolerances for grouping}
}
\value{
integer values of scan groups
}
\description{
group sequential scans if scan[i] + tolerance[i] + tolerance[i + 1] > scan[i + 1]
}
