% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{find_unique_peakgroups}
\alias{find_unique_peakgroups}
\title{Find Unique Peakgroups}
\usage{
find_unique_peakgroups(mz_set_data, peaks, rt_grouping_tol, group_corr_cutoff)
}
\arguments{
\item{mz_set_data}{a tibble containing groupIds}

\item{peaks}{a tibble containing all peaks in a mzrollDB}

\item{rt_grouping_tol}{maximum departure between retention time of two groups to be consistent}

\item{group_corr_cutoff}{minimum disagreement in terms of correlation between two groups to be consistent}
}
\value{
tibble of groupId and cluster integer
}
\description{
For a set of peaks with similar m/z group peaks which have very similar rt
and correlated abundances across samples.
}
