% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_clamr_config.R
\name{format_mass_accuracy_input}
\alias{format_mass_accuracy_input}
\title{Format Mass Accuracy Input}
\usage{
format_mass_accuracy_input(mass_accuracy_tolerance, mass_accuracy_units)
}
\arguments{
\item{mass_accuracy_tolerance}{mass accuracy of instrument in full scan data (MS1)}

\item{mass_accuracy_units}{amu or ppm}
}
\description{
Format Mass Accuracy Input
}
