% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_clamr_config.R
\name{format_mass_accuracy_tolerances}
\alias{format_mass_accuracy_tolerances}
\title{Format mass accuracy tolerances}
\usage{
format_mass_accuracy_tolerances(acquisition_params, quietly = FALSE)
}
\arguments{
\item{acquisition_params}{a named list or named vector of parameters.
Parameters which match the regular exprssion "^MS[0-9]+tol" will be formatted.
Their values should be of the form "value unit". Valid units are amu and ppm.}

\item{quietly}{Hide messages and warnings}
}
\value{
a list containing tolerances at a set of selection levels (1, 2, ...) and whether they are absolute or relative
}
\description{
Format mass accuracy tolerances
}
\examples{
acquisition_params <- list(MS1tol = "10 ppm", MS2tol = "20ppm", other_par = "bar")
acquisition_params <- c(MS1tol = "10 ppm", MS2tol = "20ppm", other_par = "bar")

}
