% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry_utils.R
\name{formula_monoisotopic_mass}
\alias{formula_monoisotopic_mass}
\title{Formula Monoisotopic Mass}
\usage{
formula_monoisotopic_mass(formulas)
}
\arguments{
\item{formulas}{a character vector of molecular formulas}
}
\value{
a tibble containing formula and exact_mass
}
\description{
Formula Monoisotopic Mass
}
\examples{
formulas <- c("C6H12O6", "C21H27N7O14P2", "C10H17N3O6S")

}
