% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakgrp_coelution_annotation.R
\name{generate_coelution_peaks}
\alias{generate_coelution_peaks}
\title{Generate Coelution Peaks}
\usage{
generate_coelution_peaks(sample_coelutions, MS1tol)
}
\arguments{
\item{sample_coelutions}{A list generated by \code{find_common_coelutions} and possibly reduced using \code{apply_coelution_labels}.}

\item{MS1tol}{from \code{\link{build_clamr_config}} or \code{\link{format_mass_accuracy_input}}}
}
\value{
A list containing:
\itemize{
  \item{binned_sample_coelutions: coelutions between unique coeluting peaks}
  \item{unique_coeluting_peaks: unique mz-scan groups for a given sample}
  }
}
\description{
Using a set of coelutions between pairs of mzs aggregate these mzs acrosss all of the coelutions that they are involved in to
create a summary of peak locations [mz, scans]
}
