% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_valid_ms_file_extensions}
\alias{get_valid_ms_file_extensions}
\title{Get all valid file extensions}
\usage{
get_valid_ms_file_extensions()
}
\description{
All valid file extensions should be stored here, and not anywhere else.
}
\details{
Anything that needs a valid file extension should call into this function.
}
