% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_similarity.R
\name{group_similar_MS2}
\alias{group_similar_MS2}
\title{Group Similar MS2}
\usage{
group_similar_MS2(mz_set, clamr_config, cosine_cutoff = 0.95)
}
\arguments{
\item{mz_set}{a set of peaks with a consistent parent mass m/z which may contain multiple species with distinct RT / MS2 fingerprints.}

\item{clamr_config}{a named list of mass spec parameters with special formatting of instrument tolerances generated by \code{\link{build_clamr_config}}.}

\item{cosine_cutoff}{Minimum cosine similarity between a pair of MS2 fragment profiles to group them into a common cluster.}
}
\description{
Wraps \code{\link{find_consistent_ms2_fingerprints}} to find consistent ms2 groups with high information
}
\examples{
clamr_config <- build_clamr_config(list(MS2tol = "20ppm"))
grouped_ms2s <- group_similar_MS2(mz_set = clamr::mz_set_example, clamr_config, cosine_cutoff = 0.95)

dplyr::count(grouped_ms2s$ms2_groups, ms2_group)

}
