% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adducts.R
\name{identify_coelutions}
\alias{identify_coelutions}
\title{Identify Coelutions}
\usage{
identify_coelutions(
  coelutions,
  type = "denovo",
  coelution_n_cutoff = 0L,
  quietly = FALSE,
  coelution_library = NULL,
  amu_tolerance = NULL,
  isotope_summary = NULL,
  ...
)
}
\arguments{
\item{coelutions}{A list containing coelutions and coelution_mz_diffs of the \code{coelution} class produced by \code{find_common_coelutions}}

\item{type}{\itemize{
 \item{\code{library}: match mzs to a list of knowns based on a specified mass accuracy tolerance -- required objects coelution_library, amu_mass_accuracy_tolerance}
 \item{\code{denovo}: match each mz to linear combinations of isotopes -- required objects: isotope_summary}
}}

\item{coelution_n_cutoff}{Minimum membership size of a mz difference group to be tested for IDs}

\item{quietly}{suppress print statements}

\item{coelution_library}{a data.frame containing the following required variables
\itemize{
 \item{\code{mz}: numeric mz differences in amu}
 \item{\code{name}: character}
 \item{\code{type}: the type of coelution event -- e.g., isotopologue, adduct, distinct_compound}
 }}

\item{amu_tolerance}{search tolerance for matches in amu -- will report the highest priority match and otherwise the closest match}

\item{isotope_summary}{isotopes that the unknown specie could be constructed with:
\itemize{
  \item{label}
  \item{amu}
  \item{count_lb (optional)}
  \item{count_ub (optional)}
  \item{prior_weight (optional)}}
If penalization = "prior", a variable named "prior" must be specified (larger weights are bad)".}

\item{...}{additional arguments to pass to coelution identification}
}
\description{
Label common mass differences using either a library of commonly seen adducts and isotopologues or optimize chemical composition as a mixed integer linear programming problem.
}
