% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{identify_split_peaks}
\alias{identify_split_peaks}
\title{Identify Split Peaks}
\usage{
identify_split_peaks(
  mzroll_db_con,
  clamr_config,
  max_rt_deviation = 5,
  ic_floor = 2^10,
  anticorrelation_co = -0.1,
  signal_frac_IQR_co = 0.75,
  spectra_corr_co = 0.8,
  stratify_regex = NULL,
  n_top_spectra_summed = 3L,
  quality_weights = c(purity = 2, quality = 1)
)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}

\item{clamr_config}{a named list of mass spec parameters with special formatting of instrument tolerances generated by \code{\link{build_clamr_config}}.}

\item{max_rt_deviation}{maximum rt deviation between two peakgroups which
could be the same analyte.}

\item{ic_floor}{floor all low or missing signals to this value for the
purpose of calculating anti-correlation in log-space.}

\item{anticorrelation_co}{cutoff for what is a strong anti-correlation as a
sign of peak splitting (most signal of some samples in group A and others in group B will induce a negative correlation between A and B).}

\item{signal_frac_IQR_co}{cutoff for interquartile range of fractional
signal spread between candidate split peak pairs.}

\item{spectra_corr_co}{cutoff for correlation between consensus spectra
of the two peakgroups.}

\item{stratify_regex}{NULL for no stratification or a string regular
expression indicating categories which should drive peak splitting
(e.g., batch or date).}

\item{n_top_spectra_summed}{integer counts of maximum number of spectra to aggregate}

\item{quality_weights}{length 2 named vector with names "purity" and "quality" indicating the relative amount to weight by precursor purity (i.e., the amount of isolated signal matching the precursorMz) versus peak quality (i.e., good peak shapes).}
}
\value{
a tibble containing two variables:
\itemize{
 \item{\code{groupId_old} - current groupIds in the mzroll_db_con}
 \item{\code{groupId_new} - updated groupIds in the mzroll_db_con}
 }
}
\description{
Find split peaks where a single analyte is distributed across one or
more peakgroups.
}
\details{
To be conservative, identified split peak pairs must satisfy all of the following conditions:
\itemize{
  \item{\code{mass agreement} - within the mass tolerance specified in the
    \code{clamr_config}}
  \item{\code{retention time agreement} - RTs are within
    \code{max_rt_deviation} of one another}
  \item{\code{mutual exclusivity} - log-abundances must be anti-correlated
    beyond \code{anticorrelation_co} and inter-quartile range of the signal
    split fracitons (fraction of signal in A versus B: A / (A + B)) above
    \code{signal_frac_IQR_co}.}
  \item{\code{batch-driven} [optional] - if batches are provided (using
    \code{stratify_regex}) then signal fraction variation should be explained
    by batches.}
  \item{\code{fragmentation agreement} - fragmentation spectra are correlated above \code{spectra_corr_co}}
}
}
