% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamr_data.R
\docType{data}
\name{isotope_summaries}
\alias{isotope_summaries}
\title{Table of isotope information, including m/z and frequency}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 289 rows and 5 columns.
}
\usage{
isotope_summaries
}
\description{
Table of isotope information, including m/z and frequency
}
\keyword{data}
