% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_searches.R
\name{join_groupIds_to_mz_rts}
\alias{join_groupIds_to_mz_rts}
\title{Join GroupIds to {mz/z, rt}}
\usage{
join_groupIds_to_mz_rts(mz_rts, mzroll_db_con, clamr_config)
}
\arguments{
\item{mz_rts}{a tibble containing mz and rt}

\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}

\item{clamr_config}{a named list of mass spec parameters with special formatting of instrument tolerances generated by \code{\link{build_clamr_config}}.}
}
\value{
mz_rts with groupIds added
}
\description{
Find the most likely groupId corresponding to a set of mz, rt tuples.
}
