% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_searches.R
\name{join_peaks_to_scans}
\alias{join_peaks_to_scans}
\title{Join peaks to scans when their is a scan range and mz match for a given sample}
\usage{
join_peaks_to_scans(mzroll_db_con)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}
}
\description{
Join peaks to scans when their is a scan range and mz match for a given sample
}
