% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_tidy_mzR.R
\name{match_mzroll_db_to_file}
\alias{match_mzroll_db_to_file}
\title{Match .mzrollDB to file}
\usage{
match_mzroll_db_to_file(mzroll_db_con, ms_file_path)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}

\item{ms_file_path}{Path to an mzXML, mzML, netCDF or mzData file}
}
\value{
a matched sampleId or NULL if no match was found
}
\description{
find the sampleId corresponding to a provided sample file path
}
