% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaksets.R
\name{match_peaksets}
\alias{match_peaksets}
\title{Match Peaksets}
\usage{
match_peaksets(query_peakset, target_peakset, cutoff = 1, all_matches = FALSE)
}
\arguments{
\item{query_peakset}{generated by \code{\link{extract_peakset}}}

\item{target_peakset}{generated by \code{\link{extract_peakset}}}

\item{cutoff}{cutoff for a valid peakset match}

\item{all_matches}{return all matches within tolerance rather than just the top match}
}
\description{
Report the best match for each query groupId (first testing for sample-level agreement of peakMz, scan, and intensity, and then testing for correlation of signals across samples)
}
