% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_utils.R
\name{merge_peakgroups}
\alias{merge_peakgroups}
\title{Merge Peakgroups}
\usage{
merge_peakgroups(mzroll_db_con, split_peaks)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}

\item{split_peaks}{output of clamr::identify_split_peaks}
}
\value{
0 invisibly (run for side effects)
}
\description{
Overwrites peakgroups and peaks tables in mzroll_db_con after merging
peakgroup pairs specified in split_peaks.
}
