% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_searches.R
\name{ms2_scans_from_raw_files}
\alias{ms2_scans_from_raw_files}
\title{MS2 Scans from files}
\usage{
ms2_scans_from_raw_files(mzroll_db_con, sample_dir = NULL)
}
\arguments{
\item{mzroll_db_con}{connection to mzroll database}

\item{sample_dir}{directory containing mzML / mzXML raw files}
}
\description{
Get all MS2 scans from all samples in a directory
}
