% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_searches.R
\name{mz_peak_join_fxn}
\alias{mz_peak_join_fxn}
\title{Match peaks to peaks for a single sample}
\usage{
mz_peak_join_fxn(sample_peaks1, sample_peaks2)
}
\arguments{
\item{sample_peaks1}{peaks for a single sample (mzmin, mzmax, minscan, maxscan)}

\item{sample_peaks2}{peaks for a single sample (mzmin, mzmax, minscan, maxscan)

Matching peaks [mzmin - mzmax, minscan - maxscan] to peaks [mzmin - mzmax, minscan - maxscan]}
}
\description{
Match peaks to peaks for a single sample
}
