% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamr_data.R
\docType{data}
\name{mz_set_example}
\alias{mz_set_example}
\title{Example data set with peak group precursor m/z, RT, and MS2 scan information.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46 rows and 6 columns.
}
\usage{
mz_set_example
}
\description{
Example data set with peak group precursor m/z, RT, and MS2 scan information.
}
\keyword{data}
