% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_mzroll_db_con.R
\name{mzroll_db_sqlite}
\alias{mzroll_db_sqlite}
\title{mzrollDB file (SQLite)}
\usage{
mzroll_db_sqlite(mzroll_db_path, unlock = TRUE)
}
\arguments{
\item{mzroll_db_path}{path to a .mzrollDB SQLite file.}

\item{unlock}{if TRUE then unlock database if database is currently locked}
}
\value{
\code{mzroll_db_con}: a connection to the mzroll/sqlite database
}
\description{
Connect to an mzrollDB SQLite database.
}
