% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaksets.R
\name{mzroll_pass_bookmarks}
\alias{mzroll_pass_bookmarks}
\title{Mzroll Pass Bookmarks}
\usage{
mzroll_pass_bookmarks(mzroll_db_path_annot, mzroll_db_path_target)
}
\arguments{
\item{mzroll_db_path_annot}{path to annotated mzroll_db generated using a subset of samples by \code{\link{reduce_mzroll_samples}}}

\item{mzroll_db_path_target}{path to destination mzroll_db with no annotations}
}
\value{
0 invisibly; function operates for its side effects
}
\description{
Mzroll Pass Bookmarks
}
