% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_search.R
\name{nest_peakgroup_features}
\alias{nest_peakgroup_features}
\title{Nest Peak Groups}
\usage{
nest_peakgroup_features(mzroll_db_con)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}
}
\value{
a tibble containing nested peaks and scans for each peak group
}
\description{
Combine the peak and scan data for each peakgroup into a 1-row per peakgroup table containing nested peak and scan lists.
}
