\name{NextPow2}
\alias{NextPow2}
\title{Next power of 2
}
\description{Find the next integer power of 2
}
\usage{
NextPow2(x)
}
\arguments{
  \item{x}{integer
}
}

\value{integer that is a power of 2 higher than
  given integer
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}


\examples{

NextPow2(600)
NextPow2(1023)
NextPow2(1025)



}
\keyword{misc}
