% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry_utils.R
\name{parse_standard_formulas}
\alias{parse_standard_formulas}
\title{Parse formulas}
\usage{
parse_standard_formulas(formulas)
}
\arguments{
\item{formulas}{a character vector of character formulas}
}
\description{
Separate chemical formulas into atom counts and return as a tibble.
}
