% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakgrp_coelution_annotation.R
\name{peakgrp_coelution_annotation}
\alias{peakgrp_coelution_annotation}
\title{Label Coelutions}
\usage{
peakgrp_coelution_annotation(
  mzroll_db_con,
  clamr_config,
  coelutions,
  minor_edge_filter_fraction = 0.01,
  min_edge_inconsistency_for_root = 1,
  root_weight_constant = 1.25
)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}

\item{clamr_config}{a named list of mass spec parameters with special formatting of instrument tolerances generated by \code{\link{build_clamr_config}}.}

\item{coelutions}{A list containing coelutions and coelution_mz_diffs of the \code{coelution} class produced by \code{find_common_coelutions}}

\item{minor_edge_filter_fraction}{peakgroup pairs which coelute less than minor_edge_filter_fraction will not be used for peakgroup aggregation"}

\item{min_edge_inconsistency_for_root}{When determining the network's roots, ignore edges represented by less than \code{min_edge_inconsistency_for_root} samples.}

\item{root_weight_constant}{W(R) = C / sum(n of daughter edges + 1)}
}
\description{
Apply labels to peak groups based on sample-level coelutions defined by mz differences characteristic of common adducts and isotope mass differences.
}
\details{
Based on set of valid mz-differences for coelutions (a coelution object containing a set of valid coelution_mz_diffs), build clusters of mzs which are linked via adducts and isotopologues.
Pairs of [mz, scan] coelutions which are aggregated if either mz-value is within the supplied mass accuracy tolerance and coelutions are within small scan difference.
Build a sample-level map of coelution relationships and then apply this to peaks.
}
