% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_mzroll_db_con.R
\name{reduce_mzroll_samples}
\alias{reduce_mzroll_samples}
\title{Reduce mzroll samples}
\usage{
reduce_mzroll_samples(mzroll_db_con, retained_sampleIds)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}

\item{retained_sampleIds}{a vector of sampleId integers}
}
\value{
update sql database return 0 invisibly
}
\description{
Remove all sample-associated features (samples, scans, peaks) which are not from the samples in retained_sampleIds
}
