% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{remove_redundant_peakgroups}
\alias{remove_redundant_peakgroups}
\title{Remove Redundant Peakgroups}
\usage{
remove_redundant_peakgroups(
  mzroll_db_con,
  clamr_config,
  rt_grouping_tol = 0.5,
  group_corr_cutoff = 0.98
)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}

\item{clamr_config}{a named list of mass spec parameters with special formatting of instrument tolerances generated by \code{\link{build_clamr_config}}.}

\item{rt_grouping_tol}{maximum departure between retention time of two groups to be consistent}

\item{group_corr_cutoff}{minimum disagreement in terms of correlation between two groups to be consistent}
}
\description{
Find and remove redundant peakgroups based on MS1 concordance Mz, Rt and Quants
}
