% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_clamr_config.R
\name{require_tolerances}
\alias{require_tolerances}
\title{CLaM R require tolerances}
\usage{
require_tolerances(clamr_config, required_msLevels)
}
\arguments{
\item{clamr_config}{a named list of mass spec parameters with special formatting of instrument tolerances generated by \code{\link{build_clamr_config}}.}

\item{required_msLevels}{an integer vector of required msLevel tolerances}
}
\value{
returns 0 invisibly
}
\description{
Require a set of tolerance parameters and verify that they are appropriately formatted.
}
