% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{smear_detection}
\alias{smear_detection}
\title{Smear Detection}
\usage{
smear_detection(
  mz_sets,
  rt_bin_cutoff_fraction = 0.5,
  rescue_peaks = TRUE,
  create_plots = TRUE
)
}
\arguments{
\item{mz_sets}{tibble containing groups of peaks with a characteristic m/z.}

\item{rt_bin_cutoff_fraction}{within a given m/z window (mz_sets), filter the window if MS2 events are in more than this fraction of the 100 RT bins; a value of 1 filters nothing.}

\item{rescue_peaks}{TRUE/FALSE; if TRUE, within "smear" m/z regions, rescue some RT intervals which contain an IC pileup relative to adjacent RT bins; if FALSE, discard all data within a "smear" m/z region.}

\item{create_plots}{TRUE/FALSE; create a plot which shows the number of RT bins and total MS2 events along with \code{rt_bin_cutoff_fraction} as a "smear" cutoff}
}
\value{
tibble which is the same as the input only with smear m/zs removed.
}
\description{
Identify and remove large background smears within a single precursor m/z.
}
