% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_matching.R
\name{spectra_filter_fragments}
\alias{spectra_filter_fragments}
\title{Spectra Filter Fragments}
\usage{
spectra_filter_fragments(fragData, minimum_ic_fraction = 0.001)
}
\arguments{
\item{fragData}{a tibble of fragmentation data such as those used in \code{\link{score_fragmentation_similarity}}.}

\item{minimum_ic_fraction}{minimum fraction of total IC for a fragment to be retained.}
}
\value{
\code{fragData} with low intensity fragments filtered.
}
\description{
Spectra Filter Fragments
}
