% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry_utils.R
\name{split_formula}
\alias{split_formula}
\title{Split formula}
\usage{
split_formula(formula, collapse = "")
}
\arguments{
\item{formula}{a chemical formula which may included losses as generated by \code{\link{create_isotopic_chemical_formula}}}

\item{collapse}{delimiter used for collapsing individual formula into single string}
}
\description{
Split a formula into individual components
}
