% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sql_get_schema}
\alias{sql_get_schema}
\title{Get SQL database schema}
\usage{
sql_get_schema(sql_con)
}
\arguments{
\item{sql_con}{a DBI connection to any sql database}
}
\description{
Get SQL database schema
}
