% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sqlite_copy}
\alias{sqlite_copy}
\title{SQLite Copy}
\usage{
sqlite_copy(sqlite_from, sqlite_to, overwrite = FALSE)
}
\arguments{
\item{sqlite_from}{an sqlite-con created by \link{create_sqlite_con}}

\item{sqlite_to}{an sqlite-con created by \link{create_sqlite_con}}

\item{overwrite}{overwrite an existing \code{sqlite_to}}
}
\value{
0 invisibly if copying occurred, 1 otherwise
}
\description{
Copy an SQLite database from one location to another, testing for
  malformity.
}
