% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamr_data.R
\docType{data}
\name{standard_coelution_library}
\alias{standard_coelution_library}
\title{Table of standard coelutions, including m/z difference and atomic description.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 39 rows and 7 columns.
}
\usage{
standard_coelution_library
}
\description{
Table of standard coelutions, including m/z difference and atomic description.
}
\keyword{data}
