% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_ms_file_extension}
\alias{strip_ms_file_extension}
\title{Strip file extension}
\usage{
strip_ms_file_extension(file_names)
}
\arguments{
\item{file_names}{list of file names. can be a mixture of types (mzML, mzXML, etc)}
}
\value{
the ms file name without extension.
}
\description{
Iterate through all valid file extensions, and remove them.
If no valid extension is found, throw an error.
}
\details{
This function is public, disallowing overwriting of valid file_extensions.
}
