% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2.R
\name{sum_spectra}
\alias{sum_spectra}
\title{Sum Spectra}
\usage{
sum_spectra(spectra_set, MS1tol)
}
\arguments{
\item{spectra_set}{a tibble containing mz, ic, and optionally label.}

\item{MS1tol}{a tolerance list created with \code{format_mass_accuracy_input}.}
}
\value{
a tibble containing mz, ic, & optionally label
}
\description{
Sum Spectra
}
