% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_clamr_config.R
\name{test_clamr_config}
\alias{test_clamr_config}
\title{Test CLaM R config}
\usage{
test_clamr_config(clamr_config)
}
\arguments{
\item{clamr_config}{a named list of mass spec parameters with special formatting of instrument tolerances generated by \code{\link{build_clamr_config}}.}
}
\description{
Test CLaM R config
}
