% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_mzroll_db_con.R
\name{test_mzroll_db_con_schema}
\alias{test_mzroll_db_con_schema}
\title{Test mzrollDB Connection Schema}
\usage{
test_mzroll_db_con_schema(mzroll_db_con)
}
\arguments{
\item{mzroll_db_con}{a connection to a mzroll database as produced by \code{\link{mzroll_db_sqlite}}}
}
\description{
Test mzrollDB Connection Schema
}
