% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_mzR_utils.R
\name{tidy_mzR_extract_mzs}
\alias{tidy_mzR_extract_mzs}
\title{Extract Masses with mzR}
\usage{
tidy_mzR_extract_mzs(
  tidy_mzR,
  clamr_config,
  mz_compounds,
  add_missing_scans = FALSE,
  plot_type = "none"
)
}
\arguments{
\item{tidy_mzR}{output of \code{\link{tidy_mzR_from_msfile}}}

\item{clamr_config}{a named list of mass spec parameters with special formatting of instrument tolerances generated by \code{\link{build_clamr_config}}.}

\item{mz_compounds}{m/z(s) to query.}

\item{add_missing_scans}{include scans with no signal as zeros in eic}

\item{plot_type}{should a plot be generated; options are: none, print and return.}
}
\value{
a list containing:
\itemize{
  \item{eic - ion counts of mz_compounds in full scan data}
  \item{fragmentation_events - meta information for MS2+ scans with mz_compounds as a precursorMz}
  \item{fragmentation_data - masses and ic of MS2+ scans}
  \item{summary_plot (optionally) - a summary plot will be added if \code{plot_type} = "return"}
  }
}
\description{
Open a specified file extract a compound of interest.
}
