% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_tidy_mzR.R
\name{tidy_mzR_from_msfile}
\alias{tidy_mzR_from_msfile}
\title{Read a mass spec file with mzR}
\usage{
tidy_mzR_from_msfile(ms_file_path, mzroll_db_path = NULL, msLevels = NULL)
}
\arguments{
\item{ms_file_path}{Path to an mzXML, mzML, netCDF or mzData file}

\item{mzroll_db_path}{path to a .mzrollDB SQLite file.}

\item{msLevels}{Filter for specific MS levels (if provided, this must be a vector of integers)}
}
\value{
tidy_mzR:
a list containing two tibbles:
\itemize{
  \item{header: summary of scan; one scan per row}
  \itemize{
    \item{acquisitionNum: scan #}
    \item{polarity: positive (1) or negative (0) mode}
    \item{other scan attributes}
  }
  \item{scan_data: abundances of ions in each scan; contains three variables: mz, ic, and scan}
  }
}
\description{
This function wraps and reformats \link[mzR]{openMSfile} converting its output to a list of tibbles.
It can open any of the file formats of openMSfile (mzXML, mzML, netCDF, mzData) as well as g-zipped versions of these formats.
}
\details{
If a \code{mzroll_db_path} is provided, then use this file's rt_update_key table (if it exists) to update retention times.
}
\seealso{
\code{\link{tidy_mzR_plot_scans}}, \code{\link{tidy_mzR_plot_major_features}} and \code{\link{tidy_mzR_extract_mzs}}
}
