% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_mzR_utils.R
\name{tidy_mzR_plot_major_features}
\alias{tidy_mzR_plot_major_features}
\title{Find Major Features}
\usage{
tidy_mzR_plot_major_features(
  tidy_mzR,
  clamr_config,
  ic_cutoff_fraction = 1/20,
  signal_cv_cutoff = 0.2
)
}
\arguments{
\item{tidy_mzR}{output of \code{\link{tidy_mzR_from_msfile}}}

\item{clamr_config}{a named list of mass spec parameters with special formatting of instrument tolerances generated by \code{\link{build_clamr_config}}.}

\item{ic_cutoff_fraction}{mzs to extract as a fraction of the maximum ic}

\item{signal_cv_cutoff}{mzs with a coefficient of variation of less than this value will be excluded since they are not eluting}
}
\value{
a ggplot2 grob
}
\description{
Load a mass spec file and find high intensity mzs
}
