% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_mzR_utils.R
\name{tidy_mzR_plot_scans}
\alias{tidy_mzR_plot_scans}
\title{Plot mzR scans}
\usage{
tidy_mzR_plot_scans(tidy_mzR, scans)
}
\arguments{
\item{tidy_mzR}{output of \code{\link{tidy_mzR_from_msfile}}}

\item{scans}{scans to plot}
}
\value{
a ggplot2 grob
}
\description{
Plot spectra for a set of scans (if more than 100 scans are provided \code{stop()} will be called).
}
