% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{unlock_sqlite}
\alias{unlock_sqlite}
\title{Unlock a locked sqlite database}
\usage{
unlock_sqlite(sqlite_con)
}
\arguments{
\item{sqlite_con}{a DBI connection to any a sqlite database}
}
\value{
an unlocked sqlite_con
}
\description{
This is a total hack, where I copy the database to /tmp and then over-write
  the original file
}
