% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{update_sample_rt}
\alias{update_sample_rt}
\title{Update Sample Retention Time}
\usage{
update_sample_rt(x, rt_var, sample_var, rt_deviation_gams)
}
\arguments{
\item{x}{an input data.frame (or data_frame)}

\item{rt_var}{the column name of an rt variable to be transformed into a new variable "rt_update"}

\item{sample_var}{a sample variable which matches the sample's retention time deviation gam}

\item{rt_deviation_gams}{a tibble of sample-specific retention time update GAMs.}
}
\description{
Using spline fits of observed retention time deviations from "average samples" update the retention time of a data.frame
}
