% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_mzroll_db_con.R
\name{validate_mzroll_db_schema}
\alias{validate_mzroll_db_schema}
\title{Validate mzrollDB Schema}
\usage{
validate_mzroll_db_schema(mzroll_data_list, tables = "all")
}
\arguments{
\item{mzroll_data_list}{a list of tables matching the mzrollDB schema or a subset of the required tables.}

\item{tables}{either "all" to compare each table to the schema and ensure that all expected tables are present or
a character vector of tables to compare.}
}
\value{
0 invisibly (used for side-effects)
}
\description{
Compare a list of tables in mzroll_data_list to the mzrollDB schema
to determine whether the list is compatible with the list schema
}
